Standardize use of send message across setup, experiments and plots
