+++ webview/src/plots/components/comparisonTable/ComparisonTableRows.tsx
@@ -1,5 +1,4 @@
 import { ComparisonPlots } from 'dvc/src/plots/webview/contract'
-import { MessageFromWebviewType } from 'dvc/src/webview/contract'
 import React, { createRef, useEffect, useState } from 'react'
 import { useDispatch } from 'react-redux'
 import { ComparisonTableColumn } from './ComparisonTableHead'
@@ -7,7 +6,7 @@ import { ComparisonTableRow } from './ComparisonTableRow'
 import { changeRowHeight, DEFAULT_ROW_HEIGHT } from './comparisonTableSlice'
 import { RowDropTarget } from './RowDropTarget'
 import { DragDropContainer } from '../../../shared/components/dragDrop/DragDropContainer'
-import { sendMessage } from '../../../shared/vscode'
+import { reorderComparisonRows } from '../../util/messages'
 
 interface ComparisonTableRowsProps {
   plots: ComparisonPlots
@@ -57,10 +56,7 @@ export const ComparisionTableRows: React.FC<ComparisonTableRowsProps> = ({
 
   const changeRowsOrder = (order: string[]) => {
     setRowsOrder(order)
-    sendMessage({
-      payload: order,
-      type: MessageFromWebviewType.REORDER_PLOTS_COMPARISON_ROWS
-    })
+    reorderComparisonRows(order)
   }
 
   return (
