+++ webview/src/plots/components/emptyState/ErrorState.tsx
@@ -1,6 +1,10 @@
 import React from 'react'
 import { EmptyState } from './EmptyState'
-import { addCustomPlot, refreshRevisions, selectRevisions } from '../messages'
+import {
+  addCustomPlot,
+  refreshRevisions,
+  selectRevisions
+} from '../../util/messages'
 import { RefreshButton } from '../../../shared/components/button/RefreshButton'
 import { StartButton } from '../../../shared/components/button/StartButton'
 import { ErrorIcon } from '../../../shared/components/errorIcon/ErrorIcon'
