+++ webview/src/setup/components/experiments/Experiments.tsx
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux'
 import { NoData } from './NoData'
 import { NeedsGitCommit } from './NeedsGitCommit'
 import { DvcLiveExamples } from './DvcLiveExamples'
-import { showExperiments } from '../messages'
+import { showExperiments } from '../../util/messages'
 import { EmptyState } from '../../../shared/components/emptyState/EmptyState'
 import { IconButton } from '../../../shared/components/button/IconButton'
 import { Beaker } from '../../../shared/components/icons'
