+++ webview/src/setup/components/studio/Settings.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useSelector } from 'react-redux'
 import { VSCodeCheckbox } from '@vscode/webview-ui-toolkit/react'
 import styles from './styles.module.scss'
-import { saveStudioToken, removeStudioToken } from '../messages'
+import { saveStudioToken, removeStudioToken } from '../../util/messages'
 import { EmptyState } from '../../../shared/components/emptyState/EmptyState'
 import { Button } from '../../../shared/components/button/Button'
 import { SetupState } from '../../store'
